% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{apply_missingness_pattern}
\alias{apply_missingness_pattern}
\title{apply_missingness_pattern}
\usage{
apply_missingness_pattern(
  x_complete,
  x_pattern,
  enforce_consecutive_missingness = FALSE
)
}
\description{
apply_missingness_pattern
}
