% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{apply_missingness_pattern_NMA}
\alias{apply_missingness_pattern_NMA}
\title{Wrapper for apply_missingness_pattern that handles NMA data}
\usage{
apply_missingness_pattern_NMA(
  x_complete,
  x_pattern,
  enforce_consecutive_missingness = FALSE,
  x_complete_missing_indicator = -999,
  x_pattern_missing_indicator = -1
)
}
\arguments{
\item{x_complete}{List with 2 matrices - complete simulated NMA data (may have rows of all -999)}

\item{x_pattern}{List with 2 matrices - pattern data}

\item{enforce_consecutive_missingness}{Passed to original function}

\item{x_complete_missing_indicator}{Value indicating "no test" in x_complete (default -999)}

\item{x_pattern_missing_indicator}{Value indicating "missing threshold" in x_pattern (default -1)}
}
\description{
Wrapper for apply_missingness_pattern that handles NMA data
}
