% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_NMA.R
\name{apply_test_indicators_to_X}
\alias{apply_test_indicators_to_X}
\title{apply_test_indicators_to_X}
\usage{
apply_test_indicators_to_X(X, indicator_matrix)
}
\description{
apply_test_indicators_to_X
}
