% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_prior_pred_checks.R
\name{basic_prior_pred_check}
\alias{basic_prior_pred_check}
\title{basic_prior_pred_check}
\usage{
basic_prior_pred_check(
  parameter = "scale",
  transform = "softplus",
  prior_mean = 0,
  prior_SD = 1,
  n_sims = 5000
)
}
\description{
basic_prior_pred_check
}
