% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots.R, R/R_fns_misc.R
\name{bc_density}
\alias{bc_density}
\title{bc_density}
\usage{
bc_density(x, lambda, observations)

bc_density(x, lambda, observations)
}
\description{
bc_density

bc_density
}
\keyword{internal}
