% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots.R, R/R_fns_misc.R
\name{box_cox_density}
\alias{box_cox_density}
\title{box_cox_density}
\usage{
box_cox_density(x, lambda, data)

box_cox_density(x, lambda, data)
}
\description{
box_cox_density

box_cox_density
}
\keyword{internal}
