% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_MA.R, R/R_fns_misc.R
\name{box_cox_grid}
\alias{box_cox_grid}
\title{box_cox_grid}
\usage{
box_cox_grid(x, lambda)

box_cox_grid(x, lambda)
}
\description{
box_cox_grid

box_cox_grid
}
\keyword{internal}
