% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{categorical_to_individual}
\alias{categorical_to_individual}
\title{categorical_to_individual}
\usage{
categorical_to_individual(
  categorical_matrix,
  binary_disease_indicator,
  missing_indicator = -1
)
}
\description{
categorical_to_individual
}
\keyword{internal}
