% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{check_if_BayesMVP_R_pkg_installed}
\alias{check_if_BayesMVP_R_pkg_installed}
\title{check_if_BayesMVP_R_pkg_installed}
\usage{
check_if_BayesMVP_R_pkg_installed(if_true = TRUE, if_false = FALSE)
}
\description{
check_if_BayesMVP_R_pkg_installed
}
\keyword{internal}
