% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{check_scalar}
\alias{check_scalar}
\title{check_scalar
Function to check scalar value}
\usage{
check_scalar(input_list, param_name)
}
\description{
check_scalar
Function to check scalar value
}
\keyword{internal}
