% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{compare_indicator_matrices}
\alias{compare_indicator_matrices}
\title{compare_indicator_matrices}
\usage{
compare_indicator_matrices(
  sim_matrix,
  real_matrix,
  test_names = c("Test1", "Test2", "Test3", "Test4"),
  detailed = TRUE
)
}
\description{
compare_indicator_matrices
}
