% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ordinal_MA.R
\name{convert_categorical_to_individual}
\alias{convert_categorical_to_individual}
\title{convert_categorical_to_individual}
\usage{
convert_categorical_to_individual(
  categorical_matrix,
  binary_disease_indicator,
  missing_indicator = -1
)
}
\description{
convert_categorical_to_individual
}
