% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ordinal_MA.R
\name{convert_cumulative_to_category}
\alias{convert_cumulative_to_category}
\title{convert_cumulative_to_category}
\usage{
convert_cumulative_to_category(cumulative_matrix, missing_indicator = -1)
}
\description{
convert_cumulative_to_category
}
