% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ordinal_MA.R
\name{convert_to_aggregate_counts_single_test}
\alias{convert_to_aggregate_counts_single_test}
\title{convert_to_aggregate_counts_single_test}
\usage{
convert_to_aggregate_counts_single_test(y_list, n_studies, n_thr)
}
\description{
convert_to_aggregate_counts_single_test
}
