% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_MA.R
\name{create_confidence_polygon}
\alias{create_confidence_polygon}
\title{create_confidence_polygon - Create credible interval polygon function}
\usage{
create_confidence_polygon(df, model_name)
}
\description{
create_confidence_polygon - Create credible interval polygon function
}
\keyword{internal}
