% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{create_cutpoint_index_matrix}
\alias{create_cutpoint_index_matrix}
\title{create_cutpoint_index_matrix}
\usage{
create_cutpoint_index_matrix(data_matrix, max_score)
}
\description{
create_cutpoint_index_matrix
}
