% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_K_fold_CV.R
\name{create_folds}
\alias{create_folds}
\title{Stratified sampling by test availability
create_folds}
\usage{
create_folds(study_test_matrix, K, seed)
}
\description{
Stratified sampling by test availability
create_folds
}
\keyword{internal}
