% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{create_indicator_index_test_in_study}
\alias{create_indicator_index_test_in_study}
\title{create_indicator_index_test_in_study}
\usage{
create_indicator_index_test_in_study(x, study_missing_indicator = -1)
}
\description{
create_indicator_index_test_in_study
}
