% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_MA.R
\name{create_prediction_polygon}
\alias{create_prediction_polygon}
\title{create_prediction_polygon -  Create prediction interval polygon function}
\usage{
create_prediction_polygon(df, model_name)
}
\description{
create_prediction_polygon -  Create prediction interval polygon function
}
\keyword{internal}
