% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_create_summary_and_traces.R
\name{create_summary_and_traces}
\alias{create_summary_and_traces}
\title{create_summary_and_traces}
\usage{
create_summary_and_traces(
  package,
  use_bridgestan,
  use_BayesMVP_for_faster_summaries,
  internal_obj,
  MCMC_params,
  model_results,
  init_object,
  n_nuisance,
  compute_main_params,
  compute_transformed_parameters,
  compute_generated_quantities,
  save_log_lik_trace,
  compute_nested_rhat,
  n_superchains,
  save_trace_tibbles
)
}
\description{
create_summary_and_traces
}
\keyword{internal}
