% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{create_test_in_study_for_NMA}
\alias{create_test_in_study_for_NMA}
\title{create_test_in_study_for_NMA - function to create "indicator_test_in_study" indicator matrix}
\usage{
create_test_in_study_for_NMA(
  seed,
  n_studies,
  n_tests_inc_ref,
  prob_present = 0.75,
  min_index_tests_per_study = 1
)
}
\description{
create_test_in_study_for_NMA - function to create "indicator_test_in_study" indicator matrix
}
