% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_for_default_args.R
\name{default_Dirichlet_random_effects_type}
\alias{default_Dirichlet_random_effects_type}
\title{default_Dirichlet_random_effects_type}
\usage{
default_Dirichlet_random_effects_type(cts, random_thresholds)
}
\description{
default_Dirichlet_random_effects_type
}
\keyword{internal}
