% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_for_default_args.R
\name{default_random_thresholds}
\alias{default_random_thresholds}
\title{default_random_thresholds}
\usage{
default_random_thresholds(cts)
}
\description{
default_random_thresholds
}
\keyword{internal}
