% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_MA.R, R/R_fns_misc.R
\name{dloglogistic}
\alias{dloglogistic}
\title{dloglogistic}
\usage{
dloglogistic(x, alpha, beta)

dloglogistic(x, alpha, beta)
}
\description{
dloglogistic

dloglogistic
}
\keyword{internal}
