% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_prep_NMA_data.R
\name{expand_test_data_to_all_studies}
\alias{expand_test_data_to_all_studies}
\title{expand_test_data_to_all_studies}
\usage{
expand_test_data_to_all_studies(
  x,
  indicator_index_test_in_study,
  study_mappings = NULL,
  missing_value = -1
)
}
\description{
expand_test_data_to_all_studies
}
\keyword{internal}
