% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_filter_params_in_string.R
\name{filter_param_draws_string_batched}
\alias{filter_param_draws_string_batched}
\title{filter_param_draws_string_batched}
\usage{
filter_param_draws_string_batched(
  named_draws_array,
  param_strings_vec,
  condition,
  exclude_NA,
  debugging = FALSE
)
}
\description{
filter_param_draws_string_batched
}
\keyword{internal}
