% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{find_all_stan_param_names}
\alias{find_all_stan_param_names}
\title{find_all_stan_param_names}
\usage{
find_all_stan_param_names(
  stan_draws_array,
  exclude_NA = FALSE,
  debugging = FALSE
)
}
\description{
find_all_stan_param_names
}
\keyword{internal}
