% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{generate_comparison_metadata}
\alias{generate_comparison_metadata}
\title{generate_comparison_metadata}
\usage{
generate_comparison_metadata(n_thr, test_names = NULL)
}
\description{
generate_comparison_metadata
}
\keyword{internal}
