% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{generate_dual_ordinal_dirichlet_priors}
\alias{generate_dual_ordinal_dirichlet_priors}
\title{generate_dual_ordinal_dirichlet_priors}
\usage{
generate_dual_ordinal_dirichlet_priors(
  n_cat,
  nd_peak = 0.15,
  d_peak = 0.3,
  nd_max_alpha = 5,
  d_max_alpha = 5,
  nd_min_alpha = 1,
  d_min_alpha = 1,
  smoothness = 0.15
)
}
\description{
generate_dual_ordinal_dirichlet_priors
}
