% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{generate_inits_for_raw_simplex_vec}
\alias{generate_inits_for_raw_simplex_vec}
\title{generate_inits_for_raw_simplex_vec}
\usage{
generate_inits_for_raw_simplex_vec(n_thr, seed, width = 0.01)
}
\description{
generate_inits_for_raw_simplex_vec
}
