% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_MA.R
\name{generate_ordered_thresholds_delta_method_correlated}
\alias{generate_ordered_thresholds_delta_method_correlated}
\title{generate_ordered_thresholds_delta_method_correlated}
\usage{
generate_ordered_thresholds_delta_method_correlated(
  mean_cutpoints,
  SD_prob_scale_vec,
  use_probit_link = TRUE,
  base_correlation = 0.5
)
}
\description{
generate_ordered_thresholds_delta_method_correlated
}
