% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_MA.R
\name{generate_ordered_thresholds_induced_dirichlet}
\alias{generate_ordered_thresholds_induced_dirichlet}
\title{generate_ordered_thresholds_induced_dirichlet}
\usage{
generate_ordered_thresholds_induced_dirichlet(
  alpha_vec,
  use_probit_link = TRUE
)
}
\description{
generate_ordered_thresholds_induced_dirichlet
}
