% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{generate_ordinal_dirichlet_prior}
\alias{generate_ordinal_dirichlet_prior}
\title{generate_ordinal_dirichlet_prior}
\usage{
generate_ordinal_dirichlet_prior(
  n_cat,
  max_alpha = 5,
  min_alpha = 1,
  left_skew_factor = 0.3
)
}
\description{
generate_ordinal_dirichlet_prior
}
