% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{generate_performance_metadata}
\alias{generate_performance_metadata}
\title{generate_performance_metadata}
\usage{
generate_performance_metadata(n_thr, test_names = NULL)
}
\description{
generate_performance_metadata
}
\keyword{internal}
