% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_generate_summary_tibble.R
\name{generate_summary_tibble}
\alias{generate_summary_tibble}
\title{generate_summary_tibble}
\usage{
generate_summary_tibble(
  n_threads = NULL,
  input_array,
  iter_dim,
  chain_dim,
  param_names,
  compute_nested_rhat,
  n_superchains
)
}
\description{
generate_summary_tibble
}
\keyword{internal}
