% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{if_null_then_set_to}
\alias{if_null_then_set_to}
\title{if_null_then_set_to}
\usage{
if_null_then_set_to(x, set_to_thif_null, debugging = FALSE)
}
\description{
if_null_then_set_to
}
\keyword{internal}
