% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_prior_pred_checks.R
\name{induced_Dirichlet_ppc_plot}
\alias{induced_Dirichlet_ppc_plot}
\title{induced_Dirichlet_ppc_plot}
\usage{
induced_Dirichlet_ppc_plot(method = "sigma", N = 5000, n_cat, other_args_list)
}
\description{
induced_Dirichlet_ppc_plot
}
