% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_Stan_model_and_json.R
\name{init_stan_model_external}
\alias{init_stan_model_external}
\title{init_stan_model_external}
\usage{
init_stan_model_external(
  debugging = FALSE,
  stan_data_list,
  stan_model_file_path,
  stan_model_obj,
  n_chains,
  init_lists_per_chain
)
}
\description{
init_stan_model_external
}
\keyword{internal}
