% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_Stan_model_and_json.R
\name{init_stan_model_internal}
\alias{init_stan_model_internal}
\title{init_stan_model_internal}
\usage{
init_stan_model_internal(
  stan_data_list,
  stan_model_name,
  stan_model_stan_include_paths = NULL,
  n_chains,
  init_lists_per_chain
)
}
\description{
init_stan_model_internal
}
\keyword{internal}
