% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_basic_helper_fns.R
\name{make_array_dims_equal_Bayesplot_order}
\alias{make_array_dims_equal_Bayesplot_order}
\title{make_array_dims_equal_Bayesplot_order}
\usage{
make_array_dims_equal_Bayesplot_order(
  input_array,
  iter_dim,
  chain_dim,
  params_dim
)
}
\description{
make_array_dims_equal_Bayesplot_order
}
\keyword{internal}
