% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots.R, R/R_fns_misc.R
\name{plot_distribution_fits}
\alias{plot_distribution_fits}
\title{plot_distribution_fits}
\usage{
plot_distribution_fits(
  individual_obs_tibble,
  n_studies,
  study_indexes = NULL,
  group_name = "non-diseased"
)

plot_distribution_fits(
  individual_obs_tibble,
  n_studies,
  study_indexes = NULL,
  group_name = "non-diseased"
)
}
\description{
plot_distribution_fits

plot_distribution_fits
}
\keyword{internal}
