% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_master_fn_prep_data_and_model.R
\name{prep_data_and_model}
\alias{prep_data_and_model}
\title{prep_data_and_model}
\usage{
prep_data_and_model(
  debugging = FALSE,
  x,
  indicator_index_test_in_study = NULL,
  internal_obj,
  basic_model_options,
  advanced_model_options,
  MCMC_params,
  priors,
  init_lists_per_chain
)
}
\description{
prep_data_and_model
}
\keyword{internal}
