% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{R_fn_apply_missingness_pattern_PHQ_9}
\alias{R_fn_apply_missingness_pattern_PHQ_9}
\title{R_fn_apply_missingness_pattern_PHQ_9}
\usage{
R_fn_apply_missingness_pattern_PHQ_9(x_PHQ9, case_list, seed = 123)
}
\description{
R_fn_apply_missingness_pattern_PHQ_9
}
