% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_misc.R
\name{R_fn_apply_thr_missingness_to_test}
\alias{R_fn_apply_thr_missingness_to_test}
\title{R_fn_apply_thr_missingness}
\usage{
R_fn_apply_thr_missingness_to_test(
  data_list_of_mat,
  thr_combo_vec = NULL,
  apply_missings_to_these_rows
)
}
\description{
R_fn_apply_thr_missingness
}
