% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_Stan_MCMC.R
\name{R_fn_compute_cmdstanr_summary_tibble}
\alias{R_fn_compute_cmdstanr_summary_tibble}
\title{R_fn_compute_cmdstanr_summary_tibble}
\usage{
R_fn_compute_cmdstanr_summary_tibble(stan_fitted_model, variables_vec)
}
\description{
R_fn_compute_cmdstanr_summary_tibble
}
\keyword{internal}
