% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_Stan_MCMC.R
\name{R_fn_compute_L_and_epsilon_from_stan_model}
\alias{R_fn_compute_L_and_epsilon_from_stan_model}
\title{R_fn_compute_L_and_epsilon_from_Stan_model}
\usage{
R_fn_compute_L_and_epsilon_from_stan_model(Min_ESS, stan_fitted_model)
}
\description{
R_fn_compute_L_and_epsilon_from_Stan_model
}
\keyword{internal}
