% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_priors_NMA.R
\name{R_fn_compute_n_thr_random_NMA}
\alias{R_fn_compute_n_thr_random_NMA}
\title{R_fn_compute_n_thr_random_NMA}
\usage{
R_fn_compute_n_thr_random_NMA(n_thr, n_studies_per_test)
}
\description{
R_fn_compute_n_thr_random_NMA
}
\keyword{internal}
