% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_prep_NMA_data.R
\name{R_fn_convert_X_to_stan_array}
\alias{R_fn_convert_X_to_stan_array}
\title{R_fn_convert_X_to_stan_array}
\usage{
R_fn_convert_X_to_stan_array(X_list)
}
\description{
R_fn_convert_X_to_stan_array
}
\keyword{internal}
