% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_create_coefficient_table}
\alias{R_fn_create_coefficient_table}
\title{Helper function to create a publication-ready table
R_fn_create_coefficient_table}
\usage{
R_fn_create_coefficient_table(
  mapped_results,
  tests_to_include = NULL,
  classes_to_include = NULL,
  covariates_to_include = NULL,
  digits = 3
)
}
\description{
Helper function to create a publication-ready table
R_fn_create_coefficient_table
}
