% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_create_dummy_variables}
\alias{R_fn_create_dummy_variables}
\title{Function to convert multi-level categorical to dummy variables
R_fn_create_dummy_variables}
\usage{
R_fn_create_dummy_variables(x, var_name, reference_level = NULL)
}
\description{
Function to convert multi-level categorical to dummy variables
R_fn_create_dummy_variables
}
