% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_create_subsetted_covariates_for_tests}
\alias{R_fn_create_subsetted_covariates_for_tests}
\title{R_fn_create_subsetted_covariates_for_tests}
\usage{
R_fn_create_subsetted_covariates_for_tests(
  cov_data,
  test_data_list,
  test_names,
  continuous_covs,
  binary_covs,
  categorical_covs,
  center_cts,
  scale_cts
)
}
\description{
R_fn_create_subsetted_covariates_for_tests
}
