% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_deviance_summary}
\alias{R_fn_deviance_summary}
\title{R_fn_deviance_summary}
\usage{
R_fn_deviance_summary(summary_tibble, test_names = NULL)
}
\description{
R_fn_deviance_summary
}
\keyword{internal}
