% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_expand_covariates_to_all_studies}
\alias{R_fn_expand_covariates_to_all_studies}
\title{R_fn_expand_covariates_to_all_studies}
\usage{
R_fn_expand_covariates_to_all_studies(
  X,
  indicator_index_test_in_study,
  study_mappings = NULL,
  missing_value = 0
)
}
\description{
R_fn_expand_covariates_to_all_studies
}
