% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_extract_NMA_AUC_summary}
\alias{R_fn_extract_NMA_AUC_summary}
\title{R_fn_extract_NMA_AUC_summary}
\usage{
R_fn_extract_NMA_AUC_summary(tibble_gq, n_thr, test_names = NULL)
}
\description{
R_fn_extract_NMA_AUC_summary
}
\keyword{internal}
