% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_extract_study_accuracy_summary}
\alias{R_fn_extract_study_accuracy_summary}
\title{R_fn_extract_study_accuracy_summary}
\usage{
R_fn_extract_study_accuracy_summary(
  summary_tibble,
  n_studies,
  n_thr,
  test_names = NULL
)
}
\description{
R_fn_extract_study_accuracy_summary
}
\keyword{internal}
