% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_generate_sroc_scenarios}
\alias{R_fn_generate_sroc_scenarios}
\title{Helper function to generate specific scenarios for sROC plots
R_fn_summarize_covariates_by_test}
\usage{
R_fn_generate_sroc_scenarios(summary_results, max_scenarios = 10)
}
\description{
Helper function to generate specific scenarios for sROC plots
R_fn_summarize_covariates_by_test
}
\keyword{internal}
