% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_generate_study_accuracy_metadata}
\alias{R_fn_generate_study_accuracy_metadata}
\title{R_fn_generate_study_accuracy_metadata}
\usage{
R_fn_generate_study_accuracy_metadata(n_studies, n_thr, test_names = NULL)
}
\description{
R_fn_generate_study_accuracy_metadata
}
\keyword{internal}
