% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_get_covariate_info_MA}
\alias{R_fn_get_covariate_info_MA}
\title{R_fn_get_covariate_info_MA}
\usage{
R_fn_get_covariate_info_MA(
  intercept_only,
  cov_data,
  model_parameterisation,
  n_studies
)
}
\description{
R_fn_get_covariate_info_MA
}
\keyword{internal}
