% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_get_covariate_info_NMA}
\alias{R_fn_get_covariate_info_NMA}
\title{R_fn_get_covariate_info_NMA}
\usage{
R_fn_get_covariate_info_NMA(
  intercept_only,
  cov_data,
  model_parameterisation,
  n_index_tests,
  n_studies
)
}
\description{
R_fn_get_covariate_info_NMA
}
\keyword{internal}
