% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_get_Stan_model_file_name.R
\name{R_fn_get_stan_model_file_name}
\alias{R_fn_get_stan_model_file_name}
\title{R_fn_get_stan_model_file_name}
\usage{
R_fn_get_stan_model_file_name(
  custom_file_name,
  cts,
  network,
  prior_only,
  model_parameterisation,
  random_thresholds,
  Dirichlet_random_effects_type
)
}
\description{
R_fn_get_stan_model_file_name
}
\keyword{internal}
