% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_GQ_covariates_NMA}
\alias{R_fn_GQ_covariates_NMA}
\title{Run standalone generated quantities for user-specified covariates}
\usage{
R_fn_GQ_covariates_NMA(
  new_cov_data,
  stan_mod_samples,
  stan_data_list,
  stan_model_file_path,
  stan_functions_directory
)
}
\description{
Run standalone generated quantities for user-specified covariates
}
