% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_Stan_model_and_json.R
\name{R_fn_init_stan_inits_external}
\alias{R_fn_init_stan_inits_external}
\title{R_fn_init_stan_inits_external}
\usage{
R_fn_init_stan_inits_external(
  stan_data_list,
  stan_model_file_path,
  stan_model_obj,
  n_chains,
  init_lists_per_chain
)
}
\description{
R_fn_init_stan_inits_external
}
\keyword{internal}
