% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_init_Stan_model_and_json.R
\name{R_fn_init_stan_inits_internal}
\alias{R_fn_init_stan_inits_internal}
\title{R_fn_init_stan_inits_internal}
\usage{
R_fn_init_stan_inits_internal(
  stan_data_list,
  stan_model_name,
  stan_model_stan_include_paths = NULL,
  n_chains_burnin,
  init_lists_per_chain
)
}
\description{
R_fn_init_stan_inits_internal
}
\keyword{internal}
