% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_log_lik_summary}
\alias{R_fn_log_lik_summary}
\title{R_fn_log_lik_summary}
\usage{
R_fn_log_lik_summary(summary_tibble, n_studies, n_thr, test_names = NULL)
}
\description{
R_fn_log_lik_summary
}
\keyword{internal}
