% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_map_beta_coefficients}
\alias{R_fn_map_beta_coefficients}
\title{R_fn_map_beta_coefficients}
\usage{
R_fn_map_beta_coefficients(
  beta_tibble,
  covariate_names = NULL,
  test_names = NULL,
  alpha = 0.05
)
}
\description{
R_fn_map_beta_coefficients
}
