% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_pad_covariate_matrices}
\alias{R_fn_pad_covariate_matrices}
\title{R_fn_pad_covariate_matrices}
\usage{
R_fn_pad_covariate_matrices(X_list, fill_value = 0)
}
\description{
R_fn_pad_covariate_matrices
}
