% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_prep_MA_data.R
\name{R_fn_prep_MA_data}
\alias{R_fn_prep_MA_data}
\title{R_fn_prep_MA_data}
\usage{
R_fn_prep_MA_data(
  x,
  cov_data,
  intercept_only,
  model_parameterisation,
  n_index_tests_per_study,
  indicator_index_test_in_study
)
}
\description{
R_fn_prep_MA_data
}
\keyword{internal}
