% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_prepare_NMA_covariates}
\alias{R_fn_prepare_NMA_covariates}
\title{R_fn_prepare_NMA_covariates}
\usage{
R_fn_prepare_NMA_covariates(
  cov_data,
  test_data_list,
  test_names,
  continuous_covs,
  binary_covs,
  categorical_covs,
  standardize_continuous = TRUE,
  handle_missing = "median",
  baseline_values = NULL
)
}
\description{
R_fn_prepare_NMA_covariates
}
