% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_K_fold_CV.R
\name{R_fn_run_kfold_cv_auto}
\alias{R_fn_run_kfold_cv_auto}
\title{R_fn_run_kfold_cv_auto
Wrapper function to choose between sequential and parallel k-fold CV}
\usage{
R_fn_run_kfold_cv_auto(
  debugging,
  K = 10,
  model_prep_obj,
  stan_data_list,
  priors,
  init_lists_per_chain,
  n_burnin,
  n_iter,
  n_chains,
  adapt_delta,
  max_treedepth,
  parallel,
  n_workers,
  output_dir,
  use_BayesMVP_for_faster_summaries
)
}
\description{
R_fn_run_kfold_cv_auto
Wrapper function to choose between sequential and parallel k-fold CV
}
