% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_K_fold_CV.R
\name{R_fn_run_kfold_cv_parallel_BayesMVP}
\alias{R_fn_run_kfold_cv_parallel_BayesMVP}
\title{R_fn_run_kfold_cv_parallel_BayesMVP
Run k-fold cross-validation in parallel across multiple folds}
\usage{
R_fn_run_kfold_cv_parallel_BayesMVP(
  debugging,
  K = 10,
  seed = 123,
  model_prep_obj,
  Stan_data_list,
  cmdstanr_args,
  priors,
  init_lists_per_chain,
  n_burnin,
  n_iter,
  n_chains_burnin,
  n_chains_sampling,
  adapt_delta,
  max_treedepth,
  n_workers = NULL,
  output_dir = "cv_results",
  use_BayesMVP_for_faster_summaries,
  stanc_args
)
}
\description{
R_fn_run_kfold_cv_parallel_BayesMVP
Run k-fold cross-validation in parallel across multiple folds
}
