% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sample.R
\name{R_fn_sample_stan_model}
\alias{R_fn_sample_stan_model}
\title{R_fn_sample_stan_model}
\usage{
R_fn_sample_stan_model(
  debugging = FALSE,
  stan_data_list,
  stan_model_obj,
  n_chains,
  init_lists_per_chain,
  n_superchains = NULL,
  seed,
  n_burnin,
  n_iter,
  adapt_delta,
  max_treedepth,
  metric_shape
)
}
\description{
R_fn_sample_stan_model
}
\keyword{internal}
