% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_priors.R
\name{R_fn_set_priors}
\alias{R_fn_set_priors}
\title{R_fn_set_priors}
\usage{
R_fn_set_priors(
  priors,
  cts,
  network,
  model_parameterisation,
  random_thresholds,
  Dirichlet_random_effects_type,
  softplus,
  n_thr,
  n_cat
)
}
\description{
R_fn_set_priors
}
\keyword{internal}
