% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_priors_MA.R
\name{R_fn_set_priors_MA}
\alias{R_fn_set_priors_MA}
\title{R_fn_set_priors_MA}
\usage{
R_fn_set_priors_MA(
  priors,
  intercept_only,
  cov_data,
  n_studies = NULL,
  n_index_tests = NULL,
  n_studies_per_test = NULL,
  cts,
  model_parameterisation,
  random_thresholds,
  Dirichlet_random_effects_type,
  softplus,
  n_thr,
  n_cat,
  indicator_index_test_in_study = NULL
)
}
\description{
R_fn_set_priors_MA
}
\keyword{internal}
