% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_priors_NMA.R
\name{R_fn_set_priors_NMA}
\alias{R_fn_set_priors_NMA}
\title{R_fn_set_priors_NMA}
\usage{
R_fn_set_priors_NMA(
  priors,
  intercept_only,
  cov_data,
  n_studies,
  n_index_tests,
  n_studies_per_test,
  cts,
  model_parameterisation,
  random_thresholds,
  Dirichlet_random_effects_type,
  softplus,
  n_thr,
  n_cat,
  indicator_index_test_in_study = NULL
)
}
\description{
R_fn_set_priors_NMA
}
\keyword{internal}
