% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_MA.R
\name{R_fn_sim_data_ord_MA}
\alias{R_fn_sim_data_ord_MA}
\title{R_fn_sim_data_ord_MA}
\usage{
R_fn_sim_data_ord_MA(
  seed = 123,
  n_studies = 50,
  N_per_study_mean = 2500,
  N_per_study_SD = 500,
  assume_perfect_GS = 1,
  true_Mean_prev = 0.2,
  true_SD_probit_prev = 0.25,
  bs_het_C_nd_prob_scale,
  bs_het_C_d_prob_scale
)
}
\description{
R_fn_sim_data_ord_MA
}
\keyword{internal}
