% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ordinal_MA.R
\name{R_fn_sim_data_ordinal_MA}
\alias{R_fn_sim_data_ordinal_MA}
\title{R_fn_sim_data_ordinal_MA}
\usage{
R_fn_sim_data_ordinal_MA(
  seed = 123,
  n_studies = 25,
  N_per_study_mean = 1000,
  N_per_study_SD = 500,
  assume_perfect_GS = 1,
  true_Mean_prev = 0.2
)
}
\description{
R_fn_sim_data_ordinal_MA
}
\keyword{internal}
