% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_NMA.R
\name{R_fn_sim_NMA_data_with_covariates_varying}
\alias{R_fn_sim_NMA_data_with_covariates_varying}
\title{R_fn_sim_NMA_data_with_covariates_varying}
\usage{
R_fn_sim_NMA_data_with_covariates_varying(
  seed = 123,
  n_studies_per_test = c(25, 20, 30, 25, 22),
  N_per_study_mean = 500,
  N_per_study_SD = 500,
  assume_perfect_GS = TRUE,
  true_Mean_prev = 0.2,
  true_SD_probit_prev = 0.25,
  simulate_covariates = TRUE,
  covariate_settings,
  original_cov_data = NULL,
  missing_value_covs = -999,
  require_index_test = TRUE
)
}
\description{
R_fn_sim_NMA_data_with_covariates_varying
}
\keyword{internal}
