% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_MA.R
\name{R_fn_sROC_plot_MA}
\alias{R_fn_sROC_plot_MA}
\title{R_fn_sROC_plot_MA}
\usage{
R_fn_sROC_plot_MA(
  stan_model_file_name,
  stan_mod_samples,
  df_true = NULL,
  conf_region_colour = "blue",
  pred_region_colour = "blue",
  n_index_tests = 1,
  n_thr = NULL,
  test_names = NULL
)
}
\description{
R_fn_sROC_plot_MA
}
\keyword{internal}
