% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots_NMA.R
\name{R_fn_sROC_plot_NMA}
\alias{R_fn_sROC_plot_NMA}
\title{R_fn_sROC_plot_NMA}
\usage{
R_fn_sROC_plot_NMA(
  stan_model_file_name,
  stan_mod_samples,
  df_true = NULL,
  conf_region_colour = "blue",
  pred_region_colour = "blue",
  n_index_tests,
  n_thr,
  test_names = NULL
)
}
\description{
R_fn_sROC_plot_NMA
}
\keyword{internal}
