% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_covariates.R
\name{R_fn_summarize_covariates_by_test}
\alias{R_fn_summarize_covariates_by_test}
\title{R_fn_summarize_covariates_by_test}
\usage{
R_fn_summarize_covariates_by_test(
  X_list,
  test_names = c("GAD-2", "GAD-7", "HADS", "BAI")
)
}
\description{
R_fn_summarize_covariates_by_test
}
\keyword{internal}
