% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_NMA_computations.R
\name{R_fn_using_Rcpp_compute_NMA_comparisons_posthoc}
\alias{R_fn_using_Rcpp_compute_NMA_comparisons_posthoc}
\title{R_fn_using_Rcpp_compute_NMA_comparisons_posthoc}
\usage{
R_fn_using_Rcpp_compute_NMA_comparisons_posthoc(
  trace_gq,
  n_thr,
  test_names = NULL
)
}
\description{
R_fn_using_Rcpp_compute_NMA_comparisons_posthoc
}
\keyword{internal}
