% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_K_fold_CV.R, R/R_fns_misc.R
\name{resize_init_list}
\alias{resize_init_list}
\title{resize_init_list}
\usage{
resize_init_list(init_lists_per_chain, n_chains_new)

resize_init_list(init_lists_per_chain, n_chains_new)
}
\description{
resize_init_list

resize_init_list
}
\keyword{internal}
