% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_MA.R
\name{SD_approx_ID_ord_prob_to_C_probit}
\alias{SD_approx_ID_ord_prob_to_C_probit}
\title{SD_approx_ID_ord_prob_to_C_probit}
\usage{
SD_approx_ID_ord_prob_to_C_probit(mean_C_cutpoint_scale, SD_prob_scale)
}
\description{
SD_approx_ID_ord_prob_to_C_probit
}
