% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ord_NMA.R
\name{simulate_covariates_preserving_missingness_FIXED}
\alias{simulate_covariates_preserving_missingness_FIXED}
\title{simulate_covariates_preserving_missingness_FIXED}
\usage{
simulate_covariates_preserving_missingness_FIXED(
  original_X,
  n_studies,
  indicator_index_test_in_study,
  covariates_same_between_classes = NULL,
  add_noise = TRUE,
  noise_sd_prop = 0.25,
  flip_prob = 0.15
)
}
\description{
simulate_covariates_preserving_missingness_FIXED
}
\keyword{internal}
