% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_sim_data_ordinal_MA.R
\name{convert_to_aggregate_counts}
\alias{convert_to_aggregate_counts}
\title{convert_to_aggregate_counts}
\usage{
convert_to_aggregate_counts(y_list, n_studies, n_tests, n_thr)
}
\description{
convert_to_aggregate_counts
}
