% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fns_for_default_args.R
\name{default_model_parameterisation}
\alias{default_model_parameterisation}
\title{default_parameterisation}
\usage{
default_model_parameterisation(cts)
}
\description{
default_parameterisation
}
\keyword{internal}
