% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_master_fn_initialise_update_run_model.R
\name{initialise_update_run_model}
\alias{initialise_update_run_model}
\title{initialise_update_run_model}
\usage{
initialise_update_run_model(
  debugging = FALSE,
  x,
  intercept_only,
  cov_data,
  indicator_index_test_in_study = NULL,
  internal_obj,
  basic_model_options,
  advanced_model_options,
  MCMC_params,
  other_advanced_options,
  priors,
  init_lists_per_chain,
  cmdstanr_args
)
}
\description{
initialise_update_run_model
}
\keyword{internal}
