% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_plot_MCMC_draws.R
\name{plot_param_group_batched}
\alias{plot_param_group_batched}
\title{plot_param_group_batched}
\usage{
plot_param_group_batched(
  draws_array,
  param_string,
  condition,
  plot_type,
  batch_size,
  debugging = FALSE
)
}
\description{
plot_param_group_batched
}
\keyword{internal}
