% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_prep_data.R
\name{R_fn_prep_MA_data}
\alias{R_fn_prep_MA_data}
\title{R_fn_prep_MA_data}
\usage{
R_fn_prep_MA_data(
  x = NULL,
  cts = NULL,
  n_thr = NULL,
  softplus = NULL,
  box_cox = NULL
)
}
\description{
R_fn_prep_MA_data
}
\keyword{internal}
