% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_prep_NMA_data.R
\name{R_fn_prep_NMA_data}
\alias{R_fn_prep_NMA_data}
\title{R_fn_prep_NMA_data}
\usage{
R_fn_prep_NMA_data(
  x,
  X,
  model_parameterisation,
  n_index_tests_per_study,
  indicator_index_test_in_study
)
}
\description{
R_fn_prep_NMA_data
}
\keyword{internal}
