% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_set_inits.R
\name{R_fn_set_inits_NMA}
\alias{R_fn_set_inits_NMA}
\title{R_fn_set_inits_NMA}
\usage{
R_fn_set_inits_NMA(
  inits,
  priors,
  n_studies,
  n_tests,
  n_thr,
  cts,
  model_parameterisation,
  random_thresholds,
  Dirichlet_random_effects_type,
  softplus
)
}
\description{
R_fn_set_inits_NMA
}
\keyword{internal}
