% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_fn_test_accuracy_plots.R
\name{R_fn_sROC_plot}
\alias{R_fn_sROC_plot}
\title{R_fn_sROC_plot}
\usage{
R_fn_sROC_plot(
  stan_model_file_name,
  stan_mod_samples,
  df_true = NULL,
  conf_region_colour = "blue",
  pred_region_colour = "blue"
)
}
\description{
R_fn_sROC_plot
}
\keyword{internal}
